Add validation in file delete API endpoint

+++ server/endpoints/v1/file/delete.ts
@@ -4,6 +4,7 @@ import Redis from "../../../database/redis";
 import Storage from "../../../storage/storage";
 import Metrics from "../../../metrics";
 import { Error } from "../../../errors";
+import Validate from "../../../validate";
 
 export default async function handleFileDelete(req: Request, match: MatchedRoute | null, ip: string | undefined): Promise<Response> {
 	if (req.method !== "DELETE") return jsonError(Error.INVALID_ENDPOINT);
@@ -21,6 +22,8 @@ export default async function handleFileDelete(req: Request, match: MatchedRoute
 	if (Number(process.env.METRICS_TYPE) >= 3) {
 		Metrics.http_auth_requests_total.labels(new URL(req.url).pathname, user).inc();
 	}
+
+	if (!Validate.userFilePathNames(data.path)) return jsonError(Error.INVALID_FILE_NAME);
 
 	let res = await Storage.deleteUserFiles(user, data.paths);
 	if (res === false) return jsonError(Error.UNKNOWN_ERROR);
